Rename and Restructure python project
**What**
- Rename to the openfaas-sdk and then restructure so that imports now
  look like `from openfaas_sdk.function.secrets import get_secret`
