+++ python/openfaas-sdk/.pylintrc
