+++ python/openfaas-sdk/LICENSE
