+++ python/openfaas-sdk/openfaas_sdk/__init__.py
@@ -1,5 +1,5 @@
 """
-function_sdk provides utility methods to help simplify and standardize
+sdk provides utility methods to help simplify and standardize
 the development of OpenFaaS functions in Python.
 """
 __version__ = "0.1.0"
