+++ python/openfaas-sdk/openfaas_sdk/function/secrets.py
