+++ python/openfaas-sdk/tests/test_function_sdk.py
@@ -1,5 +1,5 @@
 # pylint: disable-msg=C0114,C0116
-from function_sdk import __version__
+from openfaas_sdk import __version__
 
 
 def test_version():
